requirements.txt
amqp==2.5.2
asgiref==3.2.10
astroid==2.2.5
attrs==19.3.0
autobahn==20.6.2
Automat==20.2.0
billiard==3.6.2.0
celery==4.4.0
certifi==2018.11.29
cffi==1.13.2
channels==2.3.1
chardet==3.0.4
click==7.1.2
colorama==0.4.1
colour==0.1.5
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==2.5.0
Django==2.1.4
django-admin-interface==0.13.5
django-celery-results==1.2.0
django-colorfield==0.3.2
django-cors-headers==2.4.0
django-crontab==0.7.1
django-environ==0.4.5
django-extra-fields==1.0.0
django-filter==2.0.0
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-mysql==2.4.1
django-rest-knox==4.1.0
django-sslserver==0.21
djangorestframework==3.9.0
drf-yasg==1.15.0
et-xmlfile==1.0.1
geographiclib==1.49
geopy==1.19.0
h11==0.9.0
httpagentparser==1.8.1
hyperlink==19.0.0
idna==2.8
importlib-metadata==1.5.0
incremental==17.5.0
inflection==0.3.1
isort==4.3.20
itypes==1.1.0
jdcal==1.4.1
Jinja2==2.10.1
jsonschema==3.2.0
kombu==4.6.7
lazy-object-proxy==1.4.1
Markdown==3.0.1
MarkupSafe==1.1.1
mccabe==0.6.1
nameparser==1.0.2
numpy==1.16.0
openpyxl==3.0.3
pandas==1.0.3
Pillow==5.4.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.2
pylint==2.3.1
PyMySQL==0.9.3
pyodbc==4.0.30
pyOpenSSL==19.1.0
pyrsistent==0.15.7
python-dateutil==2.8.0
pytz==2018.9
redis==3.4.1
reportlab==3.5.34
requests==2.21.0
ruamel.yaml==0.15.96
service-identity==18.1.0
Shapely==1.7.0
six==1.12.0
sqlparse==0.3.1
StyleFrame==2.0.5
Twisted==20.3.0
txaio==20.4.1
typed-ast==1.4.0
uritemplate==3.0.0
urllib3==1.24.1
uvicorn==0.11.8
vine==1.3.0
websockets==8.1
wrapt==1.11.2
xlrd==1.2.0
XlsxWriter==1.2.9
zipp==3.0.0
zope.interface==5.1.0
